Implemented handlers non-existing entities

+++ pkg/repository/author.go
@@ -1,6 +1,7 @@
 package repository
 
 import (
+	"errors"
 	"fmt"
 	"github.com/jmoiron/sqlx"
 	"library-app/entities"
@@ -17,7 +18,7 @@ func NewAuthorDB(db *sqlx.DB) *AuthorDB {
 func (db *AuthorDB) CreateAuthor(author entities.AuthorCreate) (int, error) {
 	var authorId int
 	query := fmt.Sprintf("INSERT INTO %s(name, surname, description) VALUES($1, $2, $3) RETURNING author_id;", authorsTableName)
-	if err := db.QueryRow(query, author.Name, author.Surname, author.Description).Scan(&authorId); err != nil {
+	if err := db.QueryRow(query, author.Name, author.Surname, author.Description.String).Scan(&authorId); err != nil {
 		return -1, err
 	}
 	return authorId, nil
@@ -31,7 +32,6 @@ func (db *AuthorDB) GetAuthors() ([]entities.AuthorUpdate, error) {
 	if err != nil {
 		return nil, err
 	}
-	defer rows.Close()
 	for rows.Next() {
 		err := rows.StructScan(&author)
 		if err != nil {
@@ -39,26 +39,50 @@ func (db *AuthorDB) GetAuthors() ([]entities.AuthorUpdate, error) {
 		}
 		authors = append(authors, author)
 	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
 	return authors, nil
 }
 
-func (db *AuthorDB) GetAuthorById(id int) (entities.AuthorCreate, error) {
-	var author entities.AuthorCreate
-	query := fmt.Sprintf("SELECT author_id AS authorId, name, surname, description FROM %s WHERE author_id = $1", authorsTableName)
-	if err := db.Get(&author, query, id); err != nil {
-		return entities.AuthorCreate{}, err
+func (db *AuthorDB) GetAuthorById(id int) (entities.AuthorUpdate, error) {
+	if exist := db.Exist(id); exist {
+		var author entities.AuthorUpdate
+		query := fmt.Sprintf("SELECT author_id AS authorId, name, surname, description FROM %s WHERE author_id = $1", authorsTableName)
+		if err := db.Get(&author, query, id); err != nil {
+			return entities.AuthorUpdate{}, err
+		}
+		return author, nil
+	} else {
+		return entities.AuthorUpdate{}, errors.New("there is no authors with such id")
 	}
-	return author, nil
 }
 
 func (db *AuthorDB) UpdateAuthorById(id int, author entities.AuthorUpdate) error {
-	query := fmt.Sprintf("UPDATE %s SET name = $2, surname = $3, description = $4 WHERE author_id = $1", authorsTableName)
-	_, err := db.Exec(query, id, author.Name, author.Surname, author.Description.String)
-	return err
+	if exist := db.Exist(id); exist {
+		query := fmt.Sprintf("UPDATE %s SET name = $2, surname = $3, description = $4 WHERE author_id = $1", authorsTableName)
+		_, err := db.Exec(query, id, author.Name, author.Surname, author.Description.String)
+		return err
+	} else {
+		return errors.New("there is no authors with such id")
+	}
 }
 
 func (db *AuthorDB) DeleteAuthorById(id int) error {
-	query := fmt.Sprintf("DELETE FROM %s WHERE author_id = $1", authorsTableName)
-	_, err := db.Exec(query, id)
-	return err
+	if exist := db.Exist(id); exist {
+		query := fmt.Sprintf("DELETE FROM %s WHERE author_id = $1", authorsTableName)
+		_, err := db.Exec(query, id)
+		return err
+	} else {
+		return errors.New("there is no authors with such id")
+	}
+}
+
+func (db *AuthorDB) Exist(id int) bool {
+	var exist bool
+	query := "SELECT EXISTS(SELECT 1 FROM authors WHERE author_id = $1)"
+	if err := db.QueryRow(query, id).Scan(&exist); err != nil {
+		return false
+	}
+	return exist
 }
