+++ pkg/repository/genre.go
@@ -1,6 +1,7 @@
 package repository
 
 import (
+	"errors"
 	"fmt"
 	"github.com/jmoiron/sqlx"
 	"library-app/entities"
@@ -31,7 +32,6 @@ func (db *GenreDB) GetGenres() ([]entities.GenreCreate, error) {
 	if err != nil {
 		return nil, err
 	}
-	defer rows.Close()
 	for rows.Next() {
 		err := rows.StructScan(&genre)
 		if err != nil {
@@ -39,26 +39,50 @@ func (db *GenreDB) GetGenres() ([]entities.GenreCreate, error) {
 		}
 		genres = append(genres, genre)
 	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
 	return genres, nil
 }
 
 func (db *GenreDB) GetGenreById(id int) (entities.GenreCreate, error) {
-	var genre entities.GenreCreate
-	query := fmt.Sprintf("SELECT genre_id AS genreId, name FROM %s WHERE genre_id = $1", genresTableName)
-	if err := db.Get(&genre, query, id); err != nil {
-		return entities.GenreCreate{}, err
+	if exist := db.Exist(id); exist {
+		var genre entities.GenreCreate
+		query := fmt.Sprintf("SELECT genre_id AS genreId, name FROM %s WHERE genre_id = $1", genresTableName)
+		if err := db.Get(&genre, query, id); err != nil {
+			return entities.GenreCreate{}, err
+		}
+		return genre, nil
+	} else {
+		return entities.GenreCreate{}, errors.New("there is no genres with such id")
 	}
-	return genre, nil
 }
 
 func (db *GenreDB) UpdateGenreById(id int, genre entities.GenreUpdate) error {
-	query := fmt.Sprintf("UPDATE %s SET name = $2 WHERE genre_id = $1", genresTableName)
-	_, err := db.Exec(query, id, genre.Name)
-	return err
+	if exist := db.Exist(id); exist {
+		query := fmt.Sprintf("UPDATE %s SET name = $2 WHERE genre_id = $1", genresTableName)
+		_, err := db.Exec(query, id, genre.Name)
+		return err
+	} else {
+		return errors.New("there is no genres with such id")
+	}
 }
 
 func (db *GenreDB) DeleteGenreById(id int) error {
-	query := fmt.Sprintf("DELETE FROM %s WHERE genre_id = $1", genresTableName)
-	_, err := db.Exec(query, id)
-	return err
+	if exist := db.Exist(id); exist {
+		query := fmt.Sprintf("DELETE FROM %s WHERE genre_id = $1", genresTableName)
+		_, err := db.Exec(query, id)
+		return err
+	} else {
+		return errors.New("there is no genres with such id")
+	}
+}
+
+func (db *GenreDB) Exist(id int) bool {
+	var exist bool
+	query := "SELECT EXISTS(SELECT 1 FROM genres WHERE genre_id = $1)"
+	if err := db.QueryRow(query, id).Scan(&exist); err != nil {
+		return false
+	}
+	return exist
 }
